mmcglown
I'm trying to download the visual studio console 2022 so that I may create an .NET & C++ application for my windows 11 OS build MSHome computer. I have an .x32, x64 and x86 system but my "c"drive is not allowing for my SSD apps to be moved to my external internal SSD drive. I have an external Seagate that holds 10 teras of data and is not full. My x-box games are able to fit on the SSD, but the computer's internal drive is not allowing it to do so. I have a new SSD internal external drive that holds 1 tera of data and the "c" win drive will not program itself to allow me to transfer the apps over to the internal external drive so that I can free up space on my bult-in hard drive. Can you please assist me with this? I have selected the correct options for all new apps to be stored on the new SSD drive. My computer has been hacked and a governmental download sequence has been used on it without my permission. I'm an WITSEC (department of Defense Witness Protection Program) participant of the government from a political overthrow active case.